Flynn's changes from Sunday work session
 - Make versions of each autonomous mode for each chassis type.
