+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/PhatSwipeMarkerRun.java
@@ -37,67 +37,11 @@ import static org.firstinspires.ftc.robotcore.external.navigation.AxesOrder.YZX;
 import static org.firstinspires.ftc.robotcore.external.navigation.AxesReference.EXTRINSIC;
 import static org.firstinspires.ftc.robotcore.external.navigation.VuforiaLocalizer.CameraDirection.FRONT;
 
+@Autonomous(name="PhatSwipeMarkerRun", group="PhatSwipe")
+public class PhatSwipeMarkerRun extends TeamMarkerRun {
 
-/**
- */
-public abstract class BullRun4 extends StandardChassis {
-
-    private boolean madeTheRun = false;
-
-    public BullRun4(ChassisConfig config) {
-        super(config);
+    public PhatSwipeMarkerRun() {
+        super(ChassisConfig.forPhatSwipe());
     }
 
-    /**
-     * Code to run ONCE when the driver hits INIT
-     */
-    @Override
-    public void init() {
-
-        initMotors();
-    }
-
-
-    /**
-     * Code to run REPEATEDLY after the driver hits INIT, but before they hit PLAY
-     */
-    @Override
-    public void init_loop () {
-    }
-
-    /**
-     * Code to run ONCE when the driver hits PLAY
-     */
-    @Override
-    public void start () {
-        // Reset the game timer.
-        runtime.reset();
-    }
-
-    /**
-     * Code to run ONCE after the driver hits STOP
-     */
-    @Override
-    public void stop () {
-    }
-
-    /**
-     * Code to run REPEATEDLY after the driver hits PLAY but before they hit STOP
-     */
-    @Override
-    public void loop () {
-
-        if (madeTheRun == false) {
-            double speed = 0.5;
-
-            // forward 46 inches
-            encoderDrive(speed, 46, 46);
-            madeTheRun = true;
-        }
-
-        // Show the elapsed game time and wheel power.
-        telemetry.addData("Status", "time: " + runtime.toString());
-        telemetry.addData("Status", "madeTheRun=%b", madeTheRun);
-    }
 }
-
